==> Code/01b - Merge Conversations.ipynb <==
# Merge Conversations.ipynb

Combine information from all conversations into a single dictionary.

import pickle


# Files to loop over
files = [
        "Conversation_Lists_td3_gsm8k_all_others", 
        "Conversation_Lists_td3_gsm8k_tree_of_thought",
        "Conversation_Lists_td3_gsm8k_self_refine",
        "Conversation_Lists_gpt4_gsm8k_direct_prompting",
        "Conversation_Lists_gpt4_gsm8k_zero_shot_cot",
        "Conversation_Lists_gpt4_gsm8k_ape_zero_shot_cot",
        "Conversation_Lists_gpt4_gsm8k_tree_of_thought",
        "Conversation_Lists_gpt4_gsm8k_self_refine",
        "Conversation_Lists_gpt4_gsm8k_least_to_most",
        "Conversation_Lists_gpt4_gsm8k_manual_few_shot",
        "Conversation_Lists_gpt4_gsm8k_manual_cot",
        "Conversation_Lists_td3_cw_all_others",
        "Conversation_Lists_td3_cw_tree_of_thought",
        "Conversation_Lists_td3_cw_self_refine",
        "Conversation_Lists_gpt4_cw_direct_prompting",
        "Conversation_Lists_gpt4_cw_zero_shot_cot",
        "Conversation_Lists_gpt4_cw_ape_zero_shot_cot",
        "Conversation_Lists_gpt4_cw_tree_of_thought",
        "Conversation_Lists_gpt4_cw_self_refine",
        "Conversation_Lists_gpt4_cw_least_to_most",
        "Conversation_Lists_gpt4_cw_manual_few_shot",
        "Conversation_Lists_gpt4_cw_manual_cot"
         ]


# Load pickled files and merge the dictionaries
lists_dictionary = {}
for file in files:
    with open(file + '.pkl', 'rb') as f:
        lists_dictionary.update(pickle.load(f))


# Output the merged dictionary as Conversation_Lists.pkl
with open('Conversation_Lists.pkl', 'wb') as f:
    pickle.dump(lists_dictionary, f)
