==> Code/02 - Prepare for Grading.ipynb <==
# Prepare_for_Grading.ipynb

Load in conversations.

Output a text file with identifying information and conversation that will be easy to browse.

Separately, will create Excel files with the identifying info that can be used for grading and variables that need to be filled in based on the model/task/method.

import pickle

# Load conversations
with open('Conversation_Lists.pkl', 'rb') as file:
    loaded_data = pickle.load(file)

# Get dictionary items
# lists_dictionary = {'td3_gsm8k_direct_prompting_responses': td3_gsm8k_direct_prompting_responses,
#                     'td3_gsm8k_zero_shot_cot_responses': td3_gsm8k_zero_shot_cot_responses,
#                     'td3_gsm8k_ape_zero_shot_cot_responses': td3_gsm8k_ape_zero_shot_cot_responses,
#                     'td3_gsm8k_tree_of_thought_responses': td3_gsm8k_tree_of_thought_responses,
#                     'td3_gsm8k_self_refine_responses': td3_gsm8k_self_refine_responses,
#                     'td3_gsm8k_least_to_most_responses': td3_gsm8k_least_to_most_responses,
#                     'td3_gsm8k_manual_few_shot_responses': td3_gsm8k_manual_few_shot_responses,
#                     'td3_gsm8k_manual_cot_responses': td3_gsm8k_manual_cot_responses,
#                     'gpt4_gsm8k_direct_prompting_responses': gpt4_gsm8k_direct_prompting_responses,
#                     'gpt4_gsm8k_zero_shot_cot_responses': gpt4_gsm8k_zero_shot_cot_responses,
#                     'gpt4_gsm8k_ape_zero_shot_cot_responses': gpt4_gsm8k_ape_zero_shot_cot_responses,
#                     'gpt4_gsm8k_tree_of_thought_responses': gpt4_gsm8k_tree_of_thought_responses,
#                     'gpt4_gsm8k_self_refine_responses': gpt4_gsm8k_self_refine_responses,
#                     'gpt4_gsm8k_least_to_most_responses': gpt4_gsm8k_least_to_most_responses,
#                     'gpt4_gsm8k_manual_few_shot_responses': gpt4_gsm8k_manual_few_shot_responses,
#                     'gpt4_gsm8k_manual_cot_responses': gpt4_gsm8k_manual_cot_responses,
#                     'td3_cw_direct_prompting_responses': td3_cw_direct_prompting_responses,
#                     'td3_cw_zero_shot_cot_responses': td3_cw_zero_shot_cot_responses,
#                     'td3_cw_ape_zero_shot_cot_responses': td3_cw_ape_zero_shot_cot_responses,
#                     'td3_cw_tree_of_thought_responses': td3_cw_tree_of_thought_responses,
#                     'td3_cw_self_refine_responses': td3_cw_self_refine_responses,
#                     'td3_cw_least_to_most_responses': td3_cw_least_to_most_responses,
#                     'td3_cw_manual_few_shot_responses': td3_cw_manual_few_shot_responses,
#                     'td3_cw_manual_cot_responses': td3_cw_manual_cot_responses,
#                     'gpt4_cw_direct_prompting_responses': gpt4_cw_direct_prompting_responses,
#                     'gpt4_cw_zero_shot_cot_responses': gpt4_cw_zero_shot_cot_responses,
#                     'gpt4_cw_ape_zero_shot_cot_responses': gpt4_cw_ape_zero_shot_cot_responses,
#                     'gpt4_cw_tree_of_thought_responses': gpt4_cw_tree_of_thought_responses,
#                     'gpt4_cw_self_refine_responses': gpt4_cw_self_refine_responses,
#                     'gpt4_cw_least_to_most_responses': gpt4_cw_least_to_most_responses,
#                     'gpt4_cw_manual_few_shot_responses': gpt4_cw_manual_few_shot_responses,
#                     'gpt4_cw_manual_cot_responses': gpt4_cw_manual_cot_responses}
# td3_gsm8k_direct_prompting_responses = loaded_data['td3_gsm8k_direct_prompting_responses']
# td3_gsm8k_zero_shot_cot_responses = loaded_data['td3_gsm8k_zero_shot_cot_responses']
# td3_gsm8k_ape_zero_shot_cot_responses = loaded_data['td3_gsm8k_ape_zero_shot_cot_responses']
# td3_gsm8k_tree_of_thought_responses = loaded_data['td3_gsm8k_tree_of_thought_responses']
# td3_gsm8k_self_refine_responses = loaded_data['td3_gsm8k_self_refine_responses']
# td3_gsm8k_least_to_most_responses = loaded_data['td3_gsm8k_least_to_most_responses']
# td3_gsm8k_manual_few_shot_responses = loaded_data['td3_gsm8k_manual_few_shot_responses']
# td3_gsm8k_manual_cot_responses = loaded_data['td3_gsm8k_manual_cot_responses']
# gpt4_gsm8k_direct_prompting_responses = loaded_data['gpt4_gsm8k_direct_prompting_responses']
# gpt4_gsm8k_zero_shot_cot_responses = loaded_data['gpt4_gsm8k_zero_shot_cot_responses']
# gpt4_gsm8k_ape_zero_shot_cot_responses = loaded_data['gpt4_gsm8k_ape_zero_shot_cot_responses']
# gpt4_gsm8k_tree_of_thought_responses = loaded_data['gpt4_gsm8k_tree_of_thought_responses']
# gpt4_gsm8k_self_refine_responses = loaded_data['gpt4_gsm8k_self_refine_responses']
# gpt4_gsm8k_least_to_most_responses = loaded_data['gpt4_gsm8k_least_to_most_responses']
# gpt4_gsm8k_manual_few_shot_responses = loaded_data['gpt4_gsm8k_manual_few_shot_responses']
# gpt4_gsm8k_manual_cot_responses = loaded_data['gpt4_gsm8k_manual_cot_responses']
# td3_cw_direct_prompting_responses = loaded_data['td3_cw_direct_prompting_responses']
# td3_cw_zero_shot_cot_responses = loaded_data['td3_cw_zero_shot_cot_responses']
# td3_cw_ape_zero_shot_cot_responses = loaded_data['td3_cw_ape_zero_shot_cot_responses']
# td3_cw_tree_of_thought_responses = loaded_data['td3_cw_tree_of_thought_responses']
# td3_cw_self_refine_responses = loaded_data['td3_cw_self_refine_responses']
# td3_cw_least_to_most_responses = loaded_data['td3_cw_least_to_most_responses']
# td3_cw_manual_few_shot_responses = loaded_data['td3_cw_manual_few_shot_responses']
# td3_cw_manual_cot_responses = loaded_data['td3_cw_manual_cot_responses']
# gpt4_cw_direct_prompting_responses = loaded_data['gpt4_cw_direct_prompting_responses']
# gpt4_cw_zero_shot_cot_responses = loaded_data['gpt4_cw_zero_shot_cot_responses']
# gpt4_cw_ape_zero_shot_cot_responses = loaded_data['gpt4_cw_ape_zero_shot_cot_responses']
# gpt4_cw_tree_of_thought_responses = loaded_data['gpt4_cw_tree_of_thought_responses']
# gpt4_cw_self_refine_responses = loaded_data['gpt4_cw_self_refine_responses']
# gpt4_cw_least_to_most_responses = loaded_data['gpt4_cw_least_to_most_responses']
# gpt4_cw_manual_few_shot_responses = loaded_data['gpt4_cw_manual_few_shot_responses']
# gpt4_cw_manual_cot_responses = loaded_data['gpt4_cw_manual_cot_responses']


mtm_list = [mtm for mtm in loaded_data.keys()]
conversations_list = [conversation for conversation in loaded_data.values()]

print(len(mtm_list))
print(len(conversations_list))


print(mtm_list[0])
print(conversations_list[0])


# Iterate over elements of dictionary loaded_data
# Load the list at that element
# Each element in that list is a list (conversation)
# To a text file:
# Output the name of the element of loaded_data and the index of the conversation
# Output each element of the conversation on a new line
# Also save fill lists to store model_task_method and the conversation number
model_task_methods = []
conversation_numbers = []
# Delete pre-existing file
import os
if os.path.exists("conversations_for_grading.txt"):
    os.remove("conversations_for_grading.txt")
with open("conversations_for_grading.txt", "w") as file:
    for model_task_method, conversations in zip(mtm_list, conversations_list):
        #print(conversations[0])
        #print(conversations)
        #print('')
        for i, conversation in enumerate(conversations):
            #print(i)
            #print(conversation)
            #print(conversation)
            file.write(f"{model_task_method} - Conversation {i+1}\n")
            for line in conversation:
                #print(line)
                file.write(str(line) + "\n")
            file.write("\n")
            model_task_methods.append(model_task_method)
            conversation_numbers.append(i+1)


# For grading, create a dataframe of model_task_methods and conversation_numbers and output to Excel
import pandas as pd

grading_starter = pd.DataFrame({"model_task_method": model_task_methods, "conversation_number": conversation_numbers})


# GSM8K Grading
# Accuracy (0 incorrect or 1 correct), Ease of Review (1 easy to 10 hard)

# Select the model_task_method items containing gsm8k
gsm8k_grading = grading_starter[grading_starter["model_task_method"].str.contains("gsm8k")]
# Sort by model_task_method and conversation_number
#gsm8k_grading.sort_values(["model_task_method", "conversation_number"], inplace=True)

# Add columns for accuracy and ease of review
gsm8k_grading["correct"] = ""
gsm8k_grading["ease_of_review_1_easy_10_hard"] = ""

# Output to Excel
gsm8k_grading.to_excel("gsm8k_grading.xlsx", index=False)


# Creative Writing Grading
# Coherence from 1 to 10 (1 being incoherent, 10 being very coherent), whether task constraints were followed (0 not followed, 1 followed), Ease of Review (1 easy to 10 hard)

# Select the model_task_method items containing cw
cw_grading = grading_starter[grading_starter["model_task_method"].str.contains("cw")]
# Sort by model_task_method and conversation_number
#cw_grading.sort_values(["model_task_method", "conversation_number"], inplace=True)

# Add columns for coherence, task constraints followed, and ease of review
cw_grading["coherence_1_incoherent_10_very_coherent"] = ""
cw_grading["task_constraints_followed_0_not_followed_1_followed"] = ""
cw_grading["ease_of_review_1_easy_10_hard"] = ""

# Output to Excel
cw_grading.to_excel("cw_grading.xlsx", index=False)


# Ease of review grading
# Get unique model_task_methods
unique_model_task_methods = grading_starter["model_task_method"].unique()
# Put into dataframe
ease_of_review_grading = pd.DataFrame({"model_task_method": unique_model_task_methods})
# Create column ease_of_review_1_easy_10_hard
ease_of_review_grading["ease_of_review_1_easy_10_hard"] = ""
# Output as a dataframe to Excel
ease_of_review_grading.to_excel("ease_of_review_grading.xlsx", index=False)
