==> Code/03f - Make Finetuned Coherence Predictions - First 10% - Longer Prompt.ipynb <==
### Make Predictions on Withheld Data

# Whether this notebook is first 10% or last 10%
this_nb = 'first_10_LP'
# Set fine_tuned_model_id based on web interface
fine_tuned_model_id = 'ft:gpt-3.5-turbo-0613:personal:first-10-lp:8Q8p0miO'


# Get and set API key
with open('C:/Users/ijyli/Documents/OpenAI/anlp23-project.txt', 'r') as file:
    my_api_key = file.read()

#from openai import OpenAI
from openai import OpenAI
client = OpenAI(api_key=my_api_key)


# Load this_notebook's predictions
import pandas as pd
this_notebook_predictions = pd.read_csv('items_to_predict_using_' + this_nb + '.csv')


def format_test(row):

    formatted_message = [
        {
            "role": "user", 
            "content": row['response'] + "\nYour Task: Rate the coherence of the passage above on a scale of 1 to 10, 1 being incoherent and 10 being very coherent."
        }
    ]
    return formatted_message


def predict(test_messages, fine_tuned_model_id):

    response = client.chat.completions.create(
        model=fine_tuned_model_id, messages=test_messages, temperature=0, max_tokens=50
    )

    return response.choices[0].message.content


def store_predictions(test_df, fine_tuned_model_id):

    print("fine_tuned_model_id",fine_tuned_model_id)
    test_df['Prediction'] = None

    for index, row in test_df.iterrows():
        test_message = format_test(row)
        prediction_result = predict(test_message, fine_tuned_model_id)
        test_df.at[index, 'Prediction'] = prediction_result

    test_df.to_csv("predictions_based_on_" + this_nb + ".csv")


test_df = this_notebook_predictions
store_predictions(test_df, fine_tuned_model_id)
