==> Code/03k - Sentence-BERT Local Coherence.ipynb <==
# Sentence-BERT Local Coherence

Compute local coherence scores for cw responses using Sentence-BERT embeddings and average cosine similarity.

Run on Google Colab.

# sentence-transformers package setup
!pip install sentence-transformers


# Google Drive setup
from google.colab import drive
drive.mount('/content/drive')
filename_to_load = '/content/drive/My Drive/ANLP23_Project_Data/actual_cw_passage_content.xlsx'
filename_to_save = '/content/drive/My Drive/ANLP23_Project_Data/actual_cw_passage_content_with_sim.xlsx'


# Get passage content

import pandas as pd

# Read the Excel file
df = pd.read_excel(filename_to_load)

# Display the first few rows of the dataframe
print(df.head())


# Packages
from math import sqrt
from sentence_transformers import SentenceTransformer
import numpy as np
import nltk
nltk.download('punkt')
from tqdm import tqdm
tqdm.pandas()


# Load model
sentence_model = SentenceTransformer('sentence-transformers/all-distilroberta-v1')


# Testing embedding creation
embedding=sentence_model.encode("this is a sentence")
print(embedding.shape)


# Cosine similarity between embeddings
def cosine(one, two):
  return np.dot(one,two)/(sqrt(np.dot(one,one)) * sqrt(np.dot(two,two)))


# Function to get passage average cosine similarity between sentence and the following sentence
def get_passage_avg_cosine_sim(passage):
  # Split into sentences
  sentences = nltk.sent_tokenize(passage)
  # Get sentence embeddings
  sentence_embeddings = [sentence_model.encode(sentence) for sentence in sentences]
  # Saving number of sentences there were embeddings for
  num_sentences = len(sentence_embeddings)
  # Get cosine similarity between sentence and the one after it
  cosine_sim = []
  for i in range(len(sentence_embeddings)-1):
    cosine_sim.append(cosine(sentence_embeddings[i], sentence_embeddings[i+1]))
  # Return average cosine similarity
  return sum(cosine_sim)/len(cosine_sim), num_sentences


# Test function
test_1_avg, test_1_len = get_passage_avg_cosine_sim('The boat sailed out of the harbor. The captain hoped to make it halfway through the journey today.')
print('test 1')
print(test_1_avg)
print(test_1_len)

test_2_avg, test_2_len = get_passage_avg_cosine_sim('The boat sailed out of the harbor. Paul joined the other construction workers on the job site.')
print('test 2')
print(test_2_avg)
print(test_2_len)


# Create lists of average_cosine_sim, num_sentences
average_cosine_sims_responses = []
passage_num_sentences_responses = []
for response in tqdm(df['replace_slash_n_with_newline_values']):
  #print(response)
  avg, passage_num_sentences = get_passage_avg_cosine_sim(response)
  average_cosine_sims_responses.append(avg)
  passage_num_sentences_responses.append(passage_num_sentences)


# Create column for average cosine similarity
#df[['avg_cosine_sim', 'num_sentences']] = df.progress_apply(lambda row: get_passage_avg_cosine_sim(row['response']), axis=1)


# Create dataframe columns
df['avg_cosine_sim'] = average_cosine_sims_responses
df['num_sentences'] = passage_num_sentences_responses


# Write df to Excel
df.to_excel(filename_to_save)
