==> Code/06b - Coherence Score Refinement.ipynb <==
# Coherence Score Refinement

# Merge coherence score datasets
import pandas as pd
import numpy as np

cw_grading = pd.read_excel('cw_graded.xlsx')

# Limit columns but leave in response
first_10 = pd.read_csv('predictions_based_on_first_10.csv')[['model_task_method', 'conversation_number', 'Prediction', 'response']]
# Rename column Prediction to Prediction_Based_On_First_10
first_10.rename(columns={'Prediction': 'Prediction_Based_On_First_10'}, inplace=True)
# Rename column response to response_first_10
first_10.rename(columns={'response': 'response_first_10'}, inplace=True)

last_10 = pd.read_csv('predictions_based_on_last_10.csv')[['model_task_method', 'conversation_number', 'Prediction', 'response']]
# Rename column Prediction to Prediction_Based_On_Last_10
last_10.rename(columns={'Prediction': 'Prediction_Based_On_Last_10'}, inplace=True)
# Rename column response to response_last_10
last_10.rename(columns={'response': 'response_last_10'}, inplace=True)


# Merge on model_task_method and conversation_number
cw_grading_w_pred = cw_grading.merge(first_10, on=['model_task_method', 'conversation_number'], how = 'left')
cw_grading_w_pred = cw_grading_w_pred.merge(last_10, on=['model_task_method', 'conversation_number'], how = 'left')


# Create aggregated response column by taking response_first_10, response_last_10 if it is missing
cw_grading_w_pred['response'] = cw_grading_w_pred['response_first_10'].fillna(cw_grading_w_pred['response_last_10'])



# Remove {"Score": "
# Remove "}
cw_grading_w_pred['Prediction_Based_On_First_10'] = cw_grading_w_pred['Prediction_Based_On_First_10'].str.replace('{"Score": "', '')
cw_grading_w_pred['Prediction_Based_On_First_10'] = cw_grading_w_pred['Prediction_Based_On_First_10'].str.replace('"}', '')
cw_grading_w_pred['Prediction_Based_On_Last_10'] = cw_grading_w_pred['Prediction_Based_On_Last_10'].str.replace('{"Score": "', '')
cw_grading_w_pred['Prediction_Based_On_Last_10'] = cw_grading_w_pred['Prediction_Based_On_Last_10'].str.replace('"}', '')

# Convert score to float
cw_grading_w_pred['Prediction_Based_On_First_10'] = cw_grading_w_pred['Prediction_Based_On_First_10'].astype(float)
cw_grading_w_pred['Prediction_Based_On_Last_10'] = cw_grading_w_pred['Prediction_Based_On_Last_10'].astype(float)

cw_grading_w_pred


# Create mean/aggregated prediction
cw_grading_w_pred['Aggegated_Prediction'] = cw_grading_w_pred[['Prediction_Based_On_First_10', 'Prediction_Based_On_Last_10']].mean(axis=1)
cw_grading_w_pred['Aggegated_Prediction'] = cw_grading_w_pred['Aggegated_Prediction'].fillna(cw_grading_w_pred['Prediction_Based_On_First_10'])
cw_grading_w_pred['Aggegated_Prediction'] = cw_grading_w_pred['Aggegated_Prediction'].fillna(cw_grading_w_pred['Prediction_Based_On_Last_10'])


# Order columns
# model_task_method, conversation_number, response, coherence_1_incoherent_10_very_coherent, Aggregated_Prediction, diff
# Create diff, abs_diff column as difference, abs difference between coherence score and prediction
cw_grading_w_pred['diff'] = cw_grading_w_pred['coherence_1_incoherent_10_very_coherent'] - cw_grading_w_pred['Aggegated_Prediction']
cw_grading_w_pred['abs_diff'] = np.abs(cw_grading_w_pred['coherence_1_incoherent_10_very_coherent'] - cw_grading_w_pred['Aggegated_Prediction'])
cw_grading_w_pred_output = cw_grading_w_pred[['model_task_method', 'conversation_number', 'response', 'coherence_1_incoherent_10_very_coherent', 'Aggegated_Prediction', 'diff', 'abs_diff']]


# Sort by absolute value of diff descending and output to Excel
cw_grading_w_pred_output.sort_values(by='abs_diff', ascending=False, inplace=True)
cw_grading_w_pred_output.to_excel('cw_grading_w_pred.xlsx', index=False)


# Load in Combined_Data.xlsx
combined_data = pd.read_excel('Combined_Data.xlsx')


# Also create LP version
combined_data['LP_diff'] = combined_data['coherence_1_incoherent_10_very_coherent'] - combined_data['Aggregated_Prediction_LP']
combined_data['LP_abs_diff'] = np.abs(combined_data['coherence_1_incoherent_10_very_coherent'] - combined_data['Aggregated_Prediction_LP'])


# Limit columns
combined_data_output = combined_data[['model_task_method', 'conversation_number', 'response_LP', 'coherence_1_incoherent_10_very_coherent', 'Aggregated_Prediction_LP', 'LP_diff', 'LP_abs_diff']]


# Sort by absolute value of diff descending and output to Excel
combined_data_output.sort_values(by='LP_abs_diff', ascending=False, inplace=True)
combined_data_output.to_excel('cw_grading_w_pred_LP.xlsx', index=False)
