==> Code/07 - Plot and Order Settings.ipynb <==
# Order Setup

Create objects/lists of items for plot ordering.

import pandas as pd


# Release date

# Load Excel file "..\Selection of Prompt Engineering Methods\Hand-Labeled Method and Implementation Considerations.xlsx"
hand_labeled_data = pd.read_excel('..\Selection of Prompt Engineering Methods\Hand-Labeled Method and Implementation Considerations.xlsx')

# Keep columns ss_publication_date, "Prompt Engineering Method"
hand_labeled_data = hand_labeled_data[['ss_publication_date', 'Prompt Engineering Method']]

# Mapping names from this data to technqiue names
# "Few-Shot Learing" -> "manual_few_shot"
# "Chain-of-Thought Prompting" -> "manual_cot"
# "Zero-Shot Chain-of_Thought" -> "zero_shot_cot"
# "Automatic Prompt Engineer" -> "ape_zero_shot_cot"
# "Self-Refine" -> "self_refine"
# "Least-to-Most Prompting" -> "least_to_most"
# "Tree-of-Thought" -> "tree_of_thought"
# Set names in hand_labeled_data to these names
hand_labeled_data['technique_name'] = hand_labeled_data['Prompt Engineering Method'].replace({'Few-Shot Learning': 'manual_few_shot', 'Chain-of-Thought Prompting': 'manual_cot', 'Zero-Shot Chain-of-Thought': 'zero_shot_cot', 'Automatic Prompt Engineer': 'ape_zero_shot_cot', 'Self-Refine': 'self_refine', 'Least-to-Most Prompting': 'least_to_most', 'Tree-of-Thought': 'tree_of_thought'})

# Keep cases where technique_name is not 'Prompt Engineering Method'
hand_labeled_data = hand_labeled_data[hand_labeled_data['technique_name'] != hand_labeled_data['Prompt Engineering Method']]

# Convert ss_publication_date to "YYYY-MM-DD"
hand_labeled_data['ss_publication_date'] = hand_labeled_data['ss_publication_date'].dt.strftime('%Y-%m-%d')

# Drop the item with technique_name = 'tree_of_thought' that has the higher ss_publication_date
# Just keep the row with min ss_publication_date for each technique_name
hand_labeled_data = hand_labeled_data.groupby(['technique_name']).min().reset_index()
print(hand_labeled_data)


# Set method names
hand_labeled_data['Method'] = hand_labeled_data['technique_name'].replace({'ape_zero_shot_cot': 'APE Zero-Shot CoT', 'zero_shot_cot': 'Zero-Shot CoT', 'least_to_most': 'Least-to-Most', 'manual_cot': 'Manual CoT', 'manual_few_shot': 'Manual Few-Shot', 'self_refine': 'Self-Refine', 'tree_of_thought': 'Tree-of-Thought', 'direct_prompting' : 'Direct Prompting'})
print(hand_labeled_data)


# Create lists

#order_list = averages_for_changes_with_date[['Model', 'method']]
order_list = ['Text-Davinci-003', 'GPT-4']
#hue_order = list(averages_for_changes_with_date[['Method', 'ss_publication_date']].drop_duplicates().sort_values('ss_publication_date')['Method'])
hue_order_pub_date = list(hand_labeled_data[['Method', 'ss_publication_date']].drop_duplicates().sort_values('ss_publication_date')['Method'])

# Version without dp
hue_order_pub_date_no_dp = list(hand_labeled_data[['Method', 'ss_publication_date']].drop_duplicates().sort_values('ss_publication_date')['Method'])

# Add "Direct Prompting" to the end of list
hue_order_pub_date.append('Direct Prompting')

#print(hue_order)
print(hue_order_pub_date_no_dp)
print(hue_order_pub_date)
#print(hue_order_cw)


# Hue order as they were introduced in the paper
hue_order_introduced = ['Direct Prompting', 'Zero-Shot CoT', 'APE Zero-Shot CoT', 'Tree-of-Thought', 'Self-Refine', 'Least-to-Most', 'Manual Few-Shot', 'Manual CoT']

# Check contents of hue_order_introduced are the same as hue_order_gsm8k and hue_order_cw (regardless of order)
print(set(hue_order_introduced) == set(hue_order_pub_date))


# Save order_list, hue_order_gsm8k, hue_order_cw to pickle file
import pickle
with open('order_list.pkl', 'wb') as f:
    pickle.dump(order_list, f)

with open('hue_order_pub_date_no_dp.pkl', 'wb') as f:
    pickle.dump(hue_order_pub_date_no_dp, f)

with open('hue_order_pub_date.pkl', 'wb') as f:
    pickle.dump(hue_order_pub_date, f)

# with open('hue_order_gsm8k.pkl', 'wb') as f:
#     pickle.dump(hue_order_gsm8k, f)

# with open('hue_order_cw.pkl', 'wb') as f:
#     pickle.dump(hue_order_cw, f)

with open('hue_order_introduced.pkl', 'wb') as f:
    pickle.dump(hue_order_introduced, f)
