==> Code/07a - Conduct Inference.ipynb <==
# Analyze Metrics and Conduct Inference.ipynb

import pandas as pd
from statsmodels.stats.contingency_tables import mcnemar 
#import matplotlib.pyplot as plt
#import seaborn as sns
import scipy.stats as stats


# Load Combined_Data.xlsx
combined_data = pd.read_excel('Combined_Data.xlsx')

# GSM8k data - limit to task = "gsm8k"
gsm8k_data = combined_data[combined_data['task'] == 'gsm8k']

# Creative writing data - limit to task = "cw"
cw_data = combined_data[combined_data['task'] == 'cw']

# Load direct_prompting_comparison.xlsx
direct_prompting_comparison = pd.read_excel('direct_prompting_comparison.xlsx')


## Perform McNemar's Test on GSM8k data

# Get in terms of number of questions answered correctly: sum accuracy_quality column by model, method, task
gsm8k_num_correct= gsm8k_data[['model', 'method', 'task', 'correct']].groupby(['model', 'method', 'task']).sum()['correct'].reset_index()
# Rename column to num_correct
gsm8k_num_correct = gsm8k_num_correct.rename(columns={'correct': 'num_correct'})

# Make another table with counts of questions missed (number of observations minus accuracy_quality sum) by model, method, task
gsm8k_num_obs = gsm8k_data[['model', 'method', 'task', 'correct']].groupby(['model', 'method', 'task']).count()['correct'].reset_index()
# Rename column to num_obs
gsm8k_num_obs = gsm8k_num_obs.rename(columns={'correct': 'num_obs'})
# Join together by model, method, task
gsm8k_correct_incorrect_obs = gsm8k_num_correct.merge(gsm8k_num_obs, on=['model', 'method', 'task'], how='left')
# Table should have model, method, task, num_correct, num_incorrect, num_obs
gsm8k_correct_incorrect_obs['num_incorrect'] = gsm8k_correct_incorrect_obs['num_obs'] - gsm8k_correct_incorrect_obs['num_correct']

gsm8k_correct_incorrect_obs


# Function to perform McNemar's test
# Accepts argument of name of model, method, task

def perform_mcnemar(model, method, task, variable, dataframe):
    
    # Get direct prompting data
    dp_data = dataframe[(dataframe['model'] == model) & (dataframe['method'] == 'direct_prompting') & (dataframe['task'] == task)][['conversation_number', variable]]
    # Rename correct to dp_correct
    dp_data = dp_data.rename(columns={variable: 'dp_' + variable})

    # Get method data
    method_data = dataframe[(dataframe['model'] == model) & (dataframe['method'] == method) & (dataframe['task'] == task)][['conversation_number', variable]]
    # Rename correct to method_correct
    method_data = method_data.rename(columns={variable: 'method_' + variable})

    # Join together by conversation_number
    joined_data = dp_data.merge(method_data, on='conversation_number', how='left')

    # Fill contingency table cells
    dp_var_value_1_method_var_value_1 = joined_data[(joined_data['dp_' + variable] == True) & (joined_data['method_' + variable] == True)].shape[0]
    dp_var_value_1_method_var_value_2 = joined_data[(joined_data['dp_' + variable] == True) & (joined_data['method_' + variable] == False)].shape[0]
    dp_var_value_2_method_var_value_1 = joined_data[(joined_data['dp_' + variable] == False) & (joined_data['method_' + variable] == True)].shape[0]
    dp_var_value_2_method_var_value_2 = joined_data[(joined_data['dp_' + variable] == False) & (joined_data['method_' + variable] == False)].shape[0]
    
    # Create a dataset
    # Rows for dp correct and incorrect
    # Columns for method correct and incorrect
    data = [[dp_var_value_1_method_var_value_1, dp_var_value_1_method_var_value_2], 
            [dp_var_value_2_method_var_value_1, dp_var_value_2_method_var_value_2]]
  
    print('McNemar\'s Test (Exact) for ' + model + ' ' + method + ' ' + task)
    print(data)

    # McNemar's Test, exact, without any continuity correction 
    print('No continuity correction')
    print(mcnemar(data, exact=True, correction=False)) 
    ncc_result = mcnemar(data, exact=True, correction=False)
    ncc_statistic = ncc_result.statistic
    ncc_p_value = ncc_result.pvalue

    # McNemar's Test with the continuity correction 
    print('With continuity correction')
    print(mcnemar(data, exact=True, correction=True)) 
    cc_result = mcnemar(data, exact=True, correction=True)
    cc_statistic = cc_result.statistic
    cc_p_value = cc_result.pvalue

    # Return data
    return ncc_statistic, ncc_p_value, cc_statistic, cc_p_value


# Add McNemar's Test results to table
# Iterate over all combinations of model, method, task
# Except for method = "direct_prompting" and task = "cw"
# For each combination, perform McNemar's Test and add to table

# Create combos of model, method, task - all unique combinations of these three in combined_data
# Get unique values of model, method, task
models = combined_data['model'].unique()
methods = combined_data['method'].unique()
tasks = combined_data['task'].unique()
# Create list of all combinations of model, method, task
combinations = [(model, method, task) for model in models for method in methods for task in tasks]

# Add as rows to a dataframe
# Create empty dataframe
mcnemars_results = pd.DataFrame(columns=['model', 'method', 'task', 'statistic', 'pvalue', 'statistic_with_correction', 'pvalue_with_correction'])

# Iterate over combinations
for model, method, task in combinations:
    # Skip if method = "direct_prompting" or task = "cw"
    if method == 'direct_prompting' or task == 'cw':
        continue
    # Perform McNemar's Test
    statistic_without_correction, pvalue_without_correction, statistic_with_correction, pvalue_with_correction = perform_mcnemar(model, method, task, variable = 'correct', dataframe = combined_data)
    # Add to table
    mcnemars_results = pd.concat([mcnemars_results, pd.DataFrame([[model, method, task, statistic_without_correction, pvalue_without_correction, statistic_with_correction, pvalue_with_correction]], columns=['model', 'method', 'task', 'statistic', 'pvalue', 'statistic_with_correction', 'pvalue_with_correction'])])

mcnemars_results


# Rows that are significant at p < 0.05
mcnemars_results[mcnemars_results['pvalue'] < 0.05]


# Rows that are not significant at p < 0.05
mcnemars_results[mcnemars_results['pvalue'] >= 0.05]


# Printing gsm8k_correct_incorrect_obs again for reference
gsm8k_correct_incorrect_obs


# Save mcnemars_results to Excel
mcnemars_results.to_excel('mcnemars_results.xlsx', index=False)


## We can also run McNemar's test on the compliance variable

# Add as rows to a dataframe
# Create empty dataframe
mcnemars_results_compliance = pd.DataFrame(columns=['model', 'method', 'task', 'statistic', 'pvalue', 'statistic_with_correction', 'pvalue_with_correction'])

# Iterate over combinations
for model, method, task in combinations:
    # Skip if method = "direct_prompting" or task = "gsm8k"
    if method == 'direct_prompting' or task == 'gsm8k':
        continue
    # Perform McNemar's Test
    statistic_without_correction, pvalue_without_correction, statistic_with_correction, pvalue_with_correction = perform_mcnemar(model, method, task, variable = 'compliance', dataframe=combined_data)
    # Add to table
    mcnemars_results_compliance = pd.concat([mcnemars_results_compliance, pd.DataFrame([[model, method, task, statistic_without_correction, pvalue_without_correction, statistic_with_correction, pvalue_with_correction]], columns=['model', 'method', 'task', 'statistic', 'pvalue', 'statistic_with_correction', 'pvalue_with_correction'])])

mcnemars_results_compliance


# Rows that are significant at p < 0.05
mcnemars_results_compliance[mcnemars_results_compliance['pvalue'] < 0.05]


# Rows that are not significant at p < 0.05
mcnemars_results_compliance[mcnemars_results_compliance['pvalue'] >= 0.05]


# Save mcnemars_results_compliance to Excel
mcnemars_results_compliance.to_excel('mcnemars_results_compliance.xlsx', index=False)


## Paired t-test for quality, means of other metrics

print(list(direct_prompting_comparison.columns))


# Perform paired t-test on creative writing data and all other metrics

# Metrics to t-test: 'coherence_1_incoherent_10_very_coherent', 'coherence_1_incoherent_10_very_coherent_compliance_adjusted, 'ease_of_evaluation_score', 'conversation_length', 'input_length', 'output_length', 'conversation_cost', 'gsm8k_length_vs_provided', 'length_vs_direct_prompting', 'num_linebreaks', 'num_sentences', 'num_step_i', 'num_1_dot_etc', 'sentence_length', 'fres', 'num_linebreaks_prompts', 'num_sentences_prompts', 'num_step_i_prompts', 'num_1_dot_etc_prompts', 'sentence_length_prompts', 'fres_prompts'
metrics_to_t_test = ['coherence_1_incoherent_10_very_coherent', 'coherence_1_incoherent_10_very_coherent_compliance_adjusted', 'ease_of_review_1_easy_10_hard', 'conversation_length', 'input_length', 'output_length', 'conversation_cost', 'gsm8k_length_vs_provided', 'length_vs_direct_prompting', 'num_linebreaks', 'num_sentences', 'num_step_i', 'num_1_dot_etc', 'sentence_length', 'fres', 'num_linebreaks_prompts', 'num_sentences_prompts', 'num_step_i_prompts', 'num_1_dot_etc_prompts', 'sentence_length_prompts', 'fres_prompts']
# Additional metrics with cosine similarity added
additional_metrics_to_t_test = ['avg_cosine_sim', 'avg_cosine_sim_compliance_adjusted', 
'avg_inter_paragraph_cosine_sim',
'avg_inter_paragraph_cosine_sim_compliance_adjusted']
metrics_to_t_test = metrics_to_t_test + additional_metrics_to_t_test

# Create table to hold results
# Columns of model, task, method, mean metric, mean dp_metric, statistic, pvalue
t_test_results = pd.DataFrame(columns=['model', 'task', 'method', 'metric', 'dp_mean', 'using_method_mean', 'statistic', 'pvalue'])
# Iterate over model, task, method
for model, task, method in direct_prompting_comparison[['model', 'task', 'method']].drop_duplicates().values:
    # Iterate over metrics_to_t_test
    for metric in metrics_to_t_test:
        # Get data
        # Sort by conversation_number
        data = direct_prompting_comparison[(direct_prompting_comparison['model'] == model) & (direct_prompting_comparison['task'] == task) & (direct_prompting_comparison['method'] == method)].sort_values(by=['conversation_number'])
        # direct_prompting holds the metric when using direct_prompting
        direct_prompting = data['dp_' + metric]
        # using_method holds the metric when using the method
        using_method = data[metric]
        # Perform the paired sample t-test 
        statistic, pvalue = stats.ttest_rel(direct_prompting, using_method, nan_policy='omit')
        # Add to table
        t_test_results = pd.concat([t_test_results, pd.DataFrame([[model, task, method, metric, direct_prompting.mean(), using_method.mean(), statistic, pvalue]], columns=['model', 'task', 'method', 'metric', 'dp_mean', 'using_method_mean', 'statistic', 'pvalue'])])

# Significance if pvalue < 0.05
t_test_results['Significant at 95%'] = t_test_results['pvalue'].apply(lambda x: 'Yes' if x < 0.05 else 'No')

# Sort by metric, model, task, method
t_test_results = t_test_results.sort_values(by=['metric', 'model', 'task', 'method']).reset_index().drop(columns=['index'])

t_test_results


# Sort by metric, model, task, Significant at 95%
t_test_results = t_test_results.sort_values(by=['metric', 'model', 'task', 'Significant at 95%']).reset_index().drop(columns=['index'])

t_test_results


# Rows that are not significant at p < 0.05
t_test_results[t_test_results['pvalue'] >= 0.05]


# Save t_test_results to Excel
t_test_results.to_excel('t_test_results.xlsx', index=False)


## Create combined significance table

# accuracy mcnemars
# Create column Significant at 95% for pvalue < 0.05
mcnemars_results['Significant at 95%'] = mcnemars_results['pvalue_with_correction'].apply(lambda x: 'Yes' if x < 0.05 else 'No')
# Add column metric
mcnemars_results['metric'] = 'correct'

mcnemars_results


# Compliance mcnemars results
# Create column Significant at 95% for pvalue < 0.05
mcnemars_results_compliance['Significant at 95%'] = mcnemars_results_compliance['pvalue_with_correction'].apply(lambda x: 'Yes' if x < 0.05 else 'No')
# Add column metric
mcnemars_results_compliance['metric'] = 'compliance'

mcnemars_results_compliance


# All inference
all_inference = pd.concat([mcnemars_results, mcnemars_results_compliance, t_test_results], ignore_index=True)

all_inference


# Check for null values of Significant at 95%
all_inference[all_inference['Significant at 95%'].isnull()]


# Check for cases where pvalue and pvalue_with_correction are both null, but dp_mean and using_method_mean are both not null
all_inference[(all_inference['pvalue'].isnull()) & (all_inference['pvalue_with_correction'].isnull()) & (all_inference['dp_mean'].notnull()) & (all_inference['using_method_mean'].notnull())]


# Check the above but also where dp_mean and using_method_mean are not equal
all_inference[(all_inference['pvalue'].isnull()) & (all_inference['pvalue_with_correction'].isnull()) & (all_inference['dp_mean'].notnull()) & (all_inference['using_method_mean'].notnull()) & (all_inference['dp_mean'] != all_inference['using_method_mean'])]


# Output to Excel
all_inference.to_excel('all_inference.xlsx', index=False)
