==> Code/07b - Accuracy - Quality Metrics.ipynb <==
# Analyze Metrics and Conduct Inference.ipynb

Produce the large accuracy/quality table.

Metrics - GSM8K accuracy, CW Inter-Sentence Cosine Similarity, CW Inter-Paragraph Cosine Similarity, Compliance-Adjusted CW Inter-Sentence Cosine Similarity, Compliance-Adjusted CW Inter-Paragraph Cosine Similarity, Task Compliance

Row is a model by metric, column is a method.

Include ss publication date under method.

Include mean and variance (parentheses).

import pandas as pd
from statsmodels.stats.contingency_tables import mcnemar 
import matplotlib.pyplot as plt
import seaborn as sns
import scipy.stats as stats
import simpledorff
import math


# Load plot_settings.pickle
import pickle

with open('order_list.pkl', 'rb') as f:
    order_list = pickle.load(f)
print(order_list)

with open('hue_order_pub_date_no_dp.pkl', 'rb') as f:
    hue_order_pub_date_no_dp = pickle.load(f)
print(hue_order_pub_date_no_dp)

with open('hue_order_pub_date.pkl', 'rb') as f:
    hue_order_pub_date = pickle.load(f)
print(hue_order_pub_date)

with open('hue_order_introduced.pkl', 'rb') as f:
    hue_order_introduced = pickle.load(f)
print(hue_order_introduced)

# Set the font to Times New Roman or a similar serif font
import matplotlib.pyplot as plt
import matplotlib as mpl
mpl.rcParams['font.family'] = 'serif'
mpl.rcParams['font.serif'] = 'Times New Roman'
# Optionally, set the font size
mpl.rcParams['font.size'] = 12

# Figure size
f_size = (16 / 2, 9 / 2)

# Function necessary to flip the legend
import itertools
def flip(items, ncol):
    return itertools.chain(*[items[i::ncol] for i in range(ncol)])


# Load Combined_Data.xlsx
combined_data = pd.read_excel('Combined_Data.xlsx')

# GSM8k data - limit to task = "gsm8k"
gsm8k_data = combined_data[combined_data['task'] == 'gsm8k']

# Creative writing data - limit to task = "cw"
cw_data = combined_data[combined_data['task'] == 'cw']

# Load direct_prompting_comparison.xlsx
#direct_prompting_comparison = pd.read_excel('direct_prompting_comparison.xlsx')


### Plotting CW Scores for Exploration

# Loop over cases where task = "cw" and output histograms of accuracy quality by model and method
# Loop over models
for model in cw_data['model'].unique():
    # Loop over methods
    for method in cw_data['method'].unique():
        # Get data for this model and method
        data = cw_data[(cw_data['model'] == model) & (cw_data['method'] == method)]
        # Plot histogram
        plt.hist(data['coherence_1_incoherent_10_very_coherent'], bins = 10)
        # Ticks for values of 1, 2, ... 10
        plt.xticks(range(1, 11))
        plt.title(model + ' - ' + method)
        plt.xlabel('Accuracy Quality')
        plt.ylabel('Count')
        # Save histogram
        plt.savefig('../Output/cw_acc_qual_' + model + '_' + method + '.png')
        plt.close()


## Accuracy/Quality

# In combined_data, keep columns Model, Method, Task, correct, avg_cosine_sim, avg_inter_paragraph_cosine_sim, avg_cosine_sim_compliance_adjusted, avg_inter_paragraph_cosine_sim_compliance_adjusted, compliance
limited_combined_data = combined_data[['Model', 'Method', 'Task', 'model', 'method', 'task', 'correct', 'avg_cosine_sim', 'avg_inter_paragraph_cosine_sim', 'avg_cosine_sim_compliance_adjusted', 'avg_inter_paragraph_cosine_sim_compliance_adjusted', 'compliance']]

# Stack by Model and Method
stacked_combined_data = limited_combined_data.set_index(['Model', 'Method', 'Task', 'model', 'method', 'task']).stack().reset_index().rename(columns = {'level_6': 'metric', 0: 'Value'})

stacked_combined_data


# Average Value by model, method, metric
# Also get variance
avg_accuracy_quality_with_variance = stacked_combined_data.groupby(['Model', 'Method', 'Task', 'model', 'method', 'task', 'metric']).agg(['mean', 'var'])['Value'].reset_index()

# Convert variance to standard deviation
avg_accuracy_quality_with_variance['sd'] = avg_accuracy_quality_with_variance['var'].apply(math.sqrt)
# Drop the variance column
avg_accuracy_quality_with_variance.drop('var', axis = 1, inplace = True)

avg_accuracy_quality_with_variance


# Add significance
# Load all_inference.xlsx
all_inference = pd.read_excel('all_inference.xlsx')

# Merge with avg_accuracy_quality_with_variance
avg_accuracy_quality_with_variance = pd.merge(avg_accuracy_quality_with_variance, all_inference, how = 'left', on = ['model', 'method', 'task', 'metric'])

# Create column stars if Significant at 95% == "Yes"
avg_accuracy_quality_with_variance['stars'] = avg_accuracy_quality_with_variance['Significant at 95%'].apply(lambda x: '*' if x == 'Yes' else '')

avg_accuracy_quality_with_variance


# Combine mean and variance into one column that is a string with the mean and then the variance in parentheses
avg_accuracy_quality_with_variance['Value'] = avg_accuracy_quality_with_variance.apply(lambda row: str(round(row['mean'], 3)) + row['stars'] + ' (' + str(round(row['sd'], 3)) + ')', axis=1)
# But set Value to just be mean if metric is compliance or correct
avg_accuracy_quality_with_variance.loc[avg_accuracy_quality_with_variance['metric'].isin(['compliance', 'correct']), 'Value'] = avg_accuracy_quality_with_variance['mean'].round(2).astype(str) + avg_accuracy_quality_with_variance['stars']

# Drop mean and variance columns, other columns
avg_accuracy_quality_with_variance = avg_accuracy_quality_with_variance[['Model', 'Method', 'Task', 'metric', 'Value']].rename(columns = {'metric': 'Metric'})

avg_accuracy_quality_with_variance


# Pivot table - column method should go wide
avg_accuracy_quality_pivot = avg_accuracy_quality_with_variance.set_index(['Task', 'Metric', 'Model', 'Method']).unstack()
#pivot_table(index=['model', 'task'], columns='method', values='accuracy_quality').reset_index()

# Fix axis
avg_accuracy_quality_pivot = avg_accuracy_quality_pivot.rename_axis([None, None], axis=1).reset_index()

avg_accuracy_quality_pivot


# Recode Metric column
avg_accuracy_quality_pivot['Metric'] = avg_accuracy_quality_pivot['Metric'].replace({'avg_cosine_sim': 'Average Inter-Sentence Cosine Similarity', 'avg_inter_paragraph_cosine_sim': 'Average Inter-Paragraph Cosine Similarity', 'avg_cosine_sim_compliance_adjusted': 'Average Inter-Sentence Cosine Similarity (Compliance Adjusted)', 'avg_inter_paragraph_cosine_sim_compliance_adjusted': 'Average Inter-Paragraph Cosine Similarity (Compliance Adjusted)', 'compliance': 'Compliance', 'correct': 'Accuracy'})

avg_accuracy_quality_pivot


# Flatten columns
flattened_cols = [''.join(col).strip().replace('Value', '') for col in avg_accuracy_quality_pivot.columns.values]
avg_accuracy_quality_pivot.columns = flattened_cols

avg_accuracy_quality_pivot


# Print values of Metric
avg_accuracy_quality_pivot['Metric'].unique()


# Sort rows
# Task GSM8K and then Creative Writing
# Then Metric Accuracy, then Average Inter-Sentence Cosine Similarity, then Average Inter-Paragraph Cosine Similarity, then Average Inter-Sentence Cosine Similarity (Compliance Adjusted), then Average Inter-Paragraph Cosine Similarity (Compliance Adjusted), then Compliance
# Then Model Text-Davinci-003 and then GPT-4
avg_accuracy_quality_pivot['Task'] = pd.Categorical(avg_accuracy_quality_pivot['Task'], ["GSM8K", "Creative Writing"])
avg_accuracy_quality_pivot['Metric'] = pd.Categorical(avg_accuracy_quality_pivot['Metric'], ["Accuracy", "Average Inter-Sentence Cosine Similarity", "Average Inter-Paragraph Cosine Similarity", "Average Inter-Sentence Cosine Similarity (Compliance Adjusted)", "Average Inter-Paragraph Cosine Similarity (Compliance Adjusted)", "Compliance"])
avg_accuracy_quality_pivot['Model'] = pd.Categorical(avg_accuracy_quality_pivot['Model'], ["Text-Davinci-003", "GPT-4"])
avg_accuracy_quality_pivot = avg_accuracy_quality_pivot.sort_values(by=['Task', 'Metric', 'Model'])

print(avg_accuracy_quality_pivot.columns)
avg_accuracy_quality_pivot


# Order columns
starter_col_order = ['Task', 'Metric', 'Model']
col_order = starter_col_order + hue_order_pub_date
print(col_order)
avg_accuracy_quality_pivot = avg_accuracy_quality_pivot[col_order]

# Convert Model, Task, Metric back to string from Categorical
avg_accuracy_quality_pivot['Model'] = avg_accuracy_quality_pivot['Model'].astype(str)
avg_accuracy_quality_pivot['Task'] = avg_accuracy_quality_pivot['Task'].astype(str)
avg_accuracy_quality_pivot['Metric'] = avg_accuracy_quality_pivot['Metric'].astype(str)

# Delete values of Task and Metric when they are not the first instance in the dataframe
output_table = avg_accuracy_quality_pivot.copy()
output_table['Model'] = output_table['Model'].where(output_table['Model'] != output_table['Model'].shift(), '')
output_table['Task'] = output_table['Task'].where(output_table['Task'] != output_table['Task'].shift(), '')
output_table['Metric'] = output_table['Metric'].where(output_table['Metric'] != output_table['Metric'].shift(), '')

# Output to LaTeX
output_table.to_latex('../Output/avg_accuracy_quality_pivot.tex', index=False)

output_table


## Adding release date to accuracy quality table (optional version)

# Keep unique values of Method and Method + Publication Date combined_data
print(combined_data.columns)
unique_method = combined_data[['Method', 'Method + Publication Date']].drop_duplicates()

# Drop direct prompting
unique_method = unique_method[unique_method['Method'] != 'Direct Prompting']

# Convert to dictionary
unique_method = unique_method.set_index('Method').to_dict()['Method + Publication Date']

unique_method


# Use dictionary to rename columns in avg_accuracy_quality_pivot
avg_accuracy_quality_pivot_with_date = avg_accuracy_quality_pivot.rename(columns = unique_method)

print(avg_accuracy_quality_pivot_with_date.columns)

# Output to LaTeX

# Delete values of Task and Metric when they are not the first instance in the dataframe
output_table = avg_accuracy_quality_pivot_with_date.copy()
output_table['Model'] = output_table['Model'].where(output_table['Model'] != output_table['Model'].shift(), '')
output_table['Task'] = output_table['Task'].where(output_table['Task'] != output_table['Task'].shift(), '')
output_table['Metric'] = output_table['Metric'].where(output_table['Metric'] != output_table['Metric'].shift(), '')

# Output to latex. center columns, wrap text, and remove index
latex_string = output_table.to_latex(index=False, 
                      column_format='x{0.75cm}|x{1.5cm}|x{0.75cm}|x{0.75cm}|x{0.75cm}|x{0.75cm}|x{0.75cm}|x{0.75cm}|x{0.75cm}|x{0.75cm}|x{0.75cm}', 
                      #booktabs = True
                      #longtable = True
                      )

# Add lines between rows
lines = latex_string.split('\n')
new_lines = []
for line in lines:
    new_lines.append(line)
    if '\\' in line and '&' in line:  # Identifies a row of the table
        new_lines.append('\\hline')
# Insert \\hline after \toprule
new_lines.insert(2, '\\hline')

# Rejoin the modified lines
modified_latex_table = '\n'.join(new_lines)

print(modified_latex_table)

# Save string to file
with open('../Output/avg_accuracy_quality_pivot_with_date.tex', 'w') as f:
    f.write(modified_latex_table)

output_table
