==> Code/08 - Run Regressions.ipynb <==
# Run Regressions.ipynb

Logistic regressions of GSM8K accuracy on length and complexity variables, quadratic terms, clustering by identity of the question and method (task, conversation_id, method).

Regression of CW scores similarly.

Add complexity of provided answers as an interaction term in GSM8K regressions.

Add model as an interaction term.

Model and question (conversation_number by task) controls can soak up additional variation.

import statsmodels.api as sm
import pandas as pd
import statsmodels.formula.api as smf


# Load Combined_Data.xlsx
df = pd.read_excel('../Data/Combined_Data.xlsx')

print(df.columns)

df


# Transformation - divide conversation length by 1000 to get effect per 1000K tokens
df['conversation_length_thousands'] = df['conversation_length']/1000


# Create quadratics
df['conversation_length_thousands_2'] = df['conversation_length_thousands']**2
df['consolidated_num_steps_ideas_2'] = df['consolidated_num_steps_ideas']**2
df['fres_2'] = df['fres']**2


# Create clustering variable
# Concatenate task, conversation_number, method
df['task_conversation_method'] = df['task'].astype(str) + "_" + df['conversation_number'].astype(str) + "_" + df['method'].astype(str)


# Task by conversation variable as a control
df['task_conversation'] = df['task'].astype(str) + "_" + df['conversation_number'].astype(str)


# Split data
gsm8k_data = df[df['task'] == 'gsm8k']
cw_data = df[df['task'] == 'cw']


### Some Checks

# Get values of correct in gsm8k_data
gsm8k_correct = gsm8k_data['correct'].values
print(set(gsm8k_correct))

# Print cases where correct is not 0 or 1
#print(gsm8k_data[gsm8k_data['correct'] != 0 & gsm8k_data['correct'] != 1])


### Functions for table creation

# Create logit df
def create_logit_results_df(model, marginal_effects, title):
    #coef_names = model.params.index
    #print(coef_names)
    coef_names_with_mes = list(model.params.index)[1:]
    #print(coef_names_with_mes)
    me_values = marginal_effects.margeff
    #print(me_values)
    me_ses = marginal_effects.margeff_se
    # print(me_ses)
    p_values = marginal_effects.pvalues
    # print(p_values)
    # Loop over items in coef_names_with_mes, enumerated
    string_entries = []
    for i, coef_name in enumerate(coef_names_with_mes):
        # Print coef_name, me_values[i], me_ses[i], p_values[i]
        #print(coef_name, me_values[i], me_ses[i], p_values[i])
        # Consolidate me_values[i], me_ses[i], p_values[i] into a string
        # me_value* (me_se), where the star is if p_value < 0.05
        if p_values[i] < 0.05:
            string_entries.append(str(round(me_values[i], 3)) + "* (" + str(round(me_ses[i], 3)) + ")")
        else:
            string_entries.append(str(round(me_values[i], 3)) + " (" + str(round(me_ses[i], 3)) + ")")

    # Make a dataframe with one row with columns of coef_names_with_mes and values of string_entries
    # Print the dataframe
    model_df = pd.DataFrame([string_entries], columns=coef_names_with_mes)
    # Add column Title
    model_df['Title'] = title
    return model_df

# Create linear regression df
def create_linear_results_df(model, title):
    coef_names = model.params.index
    # print(coef_names)
    coef_values = model.params.values
    # print(coef_values)
    sds = model.bse.values
    # print(sds)
    p_values = model.pvalues.values
    # print(p_values)

    # Loop over items in coef_names, enumerated
    string_entries = []
    for i, coef_name in enumerate(coef_names):
        # Print coef_name, coef_values[i], sds[i], p_values[i]
        #print(coef_name, coef_values[i], sds[i], p_values[i])
        # Consolidate coef_values[i], sds[i], p_values[i] into a string
        # coef_value* (sd), where the star is if p_value < 0.05
        if p_values[i] < 0.05:
            string_entries.append(str(round(coef_values[i], 3)) + "* (" + str(round(sds[i], 3)) + ")")
        else:
            string_entries.append(str(round(coef_values[i], 3)) + " (" + str(round(sds[i], 3)) + ")")

    # Make a dataframe with one row with columns of coef_names and values of string_entries
    # Print the dataframe
    model_df = pd.DataFrame([string_entries], columns=coef_names)
    # Add column Title
    model_df['Title'] = title
    #print(lpm_with_clustering_gsm8k_df)
    return model_df



### GSM8K Regressions

#### Logistic Regression (no clustering)

# Define the logistic regression model[T.td3][T.td3][T.td3]
logit_no_clustering_gsm8k = smf.logit('correct ~ conversation_length_thousands + consolidated_num_steps_ideas + conversation_length_thousands_2 + consolidated_num_steps_ideas_2 + model', data=gsm8k_data).fit(cov_type='HC3')

# Display the summary
print(logit_no_clustering_gsm8k.summary())

# Marginal effects
logit_no_clustering_gsm8k_marginal_effects = logit_no_clustering_gsm8k.get_margeff(at='overall')
print(logit_no_clustering_gsm8k_marginal_effects.summary())

# Use function to create results df
logit_no_clustering_gsm8k_df = create_logit_results_df(logit_no_clustering_gsm8k, logit_no_clustering_gsm8k_marginal_effects, "GSM8K Correct, Logit")
print(logit_no_clustering_gsm8k_df)


#### Linear Probability Model with Clustering

# Define and fit the OLS model with clustered standard errors
lpm_with_clustering_gsm8k = smf.ols('correct ~ conversation_length_thousands + consolidated_num_steps_ideas + conversation_length_thousands_2 + consolidated_num_steps_ideas_2 + task_conversation + model', data=gsm8k_data).fit(cov_type='cluster', cov_kwds={'groups': gsm8k_data['task_conversation_method']})

# Print the model summary
print(lpm_with_clustering_gsm8k.summary())

# # Results for table
# # Get conversation_length_thousands coefficient, star for significant, sd
# # Also get consolidated_num_steps_ideas coefficient, star for significant, sd
# # Also get conversation_length_thousands_2 coefficient, star for significant, sd
# # Also get consolidated_num_steps_ideas_2 coefficient, star for significant, sd
# # Title: "GSM8K Correct, Linear"
# coef_names = lpm_with_clustering_gsm8k.params.index
# # print(coef_names)
# coef_values = lpm_with_clustering_gsm8k.params.values
# # print(coef_values)
# sds = lpm_with_clustering_gsm8k.bse.values
# # print(sds)
# p_values = lpm_with_clustering_gsm8k.pvalues.values
# # print(p_values)

# # Loop over items in coef_names, enumerated
# string_entries = []
# for i, coef_name in enumerate(coef_names):
#     # Print coef_name, coef_values[i], sds[i], p_values[i]
#     #print(coef_name, coef_values[i], sds[i], p_values[i])
#     # Consolidate coef_values[i], sds[i], p_values[i] into a string
#     # coef_value* (sd), where the star is if p_value < 0.05
#     if p_values[i] < 0.05:
#         string_entries.append(str(round(coef_values[i], 3)) + "* (" + str(round(sds[i], 3)) + ")")
#     else:
#         string_entries.append(str(round(coef_values[i], 3)) + " (" + str(round(sds[i], 3)) + ")")

# # Make a dataframe with one row with columns of coef_names and values of string_entries
# # Print the dataframe
# lpm_with_clustering_gsm8k_df = pd.DataFrame([string_entries], columns=coef_names)
# # Add column Title
# lpm_with_clustering_gsm8k_df['Title'] = "GSM8K Correct, Linear"
# print(lpm_with_clustering_gsm8k_df)

# Use function to create results df
lpm_with_clustering_gsm8k_df = create_linear_results_df(lpm_with_clustering_gsm8k, "GSM8K Correct, Linear")
print(lpm_with_clustering_gsm8k_df)


### CW Regressions

# Check
# Is avg_inter_paragraph_cosine_sim always present
#print(cw_data[cw_data['avg_inter_paragraph_cosine_sim'].isnull()])

# Check length
#print(len(cw_data['avg_inter_paragraph_cosine_sim']))
#print(len(cw_data['fres'].dropna()))

# Limit cw_data to rows where avg_inter_paragraph_cosine_sim is not null
cw_data = cw_data[cw_data['avg_inter_paragraph_cosine_sim'].notnull()]


#### Preferred cosine similarity measure

# Define and fit the OLS model with clustered standard errors
reg_with_clustering_cw = smf.ols("avg_inter_paragraph_cosine_sim ~ conversation_length_thousands + consolidated_num_steps_ideas + conversation_length_thousands_2 + consolidated_num_steps_ideas_2 + fres + fres_2 + task_conversation + model",
                                 data=cw_data).fit(cov_type='cluster', cov_kwds={'groups': cw_data['task_conversation_method']})

# Print the model summary
print(reg_with_clustering_cw.summary())

# Use function to create results df
reg_with_clustering_cw_df = create_linear_results_df(reg_with_clustering_cw, "Creative Writing Cosine Similarity")
print(reg_with_clustering_cw_df)


#### Check task compliance as well

print(list(cw_data.columns))


# Define the logistic regression model
logit_no_clustering_cw_compliance = smf.logit('compliance ~ conversation_length_thousands + consolidated_num_steps_ideas + conversation_length_thousands_2 + consolidated_num_steps_ideas_2 + fres + fres_2 + model', data=cw_data).fit(cov_type='HC3')

print('logit no clustering cw compliance')
# Display the summary
print(logit_no_clustering_cw_compliance.summary())

# Marginal effects
logit_no_clustering_cw_compliance_marginal_effects = logit_no_clustering_cw_compliance.get_margeff(at='overall')
print(logit_no_clustering_cw_compliance_marginal_effects.summary())

# Use function to create results df
logit_no_clustering_cw_compliance_df = create_logit_results_df(logit_no_clustering_cw_compliance, logit_no_clustering_cw_compliance_marginal_effects, "Creative Writing Compliance, Logit")
print(logit_no_clustering_cw_compliance_df)


# Define and fit the OLS model with clustered standard errors
lpm_with_clustering_cw_compliance = smf.ols('compliance ~ conversation_length_thousands + consolidated_num_steps_ideas + conversation_length_thousands_2 + consolidated_num_steps_ideas_2 + fres + fres_2 + task_conversation + model', data=cw_data).fit(cov_type='cluster', cov_kwds={'groups': cw_data['task_conversation_method']})

# Print the model summary
print('lpm with clustering cw compliance')
print(lpm_with_clustering_cw_compliance.summary())

# Use function to create results df
lpm_with_clustering_cw_compliance_df = create_linear_results_df(lpm_with_clustering_cw_compliance, "Creative Writing Compliance, Linear")
print(lpm_with_clustering_cw_compliance_df)


# Stack results dataframes together for output
results_df = pd.concat([logit_no_clustering_gsm8k_df, lpm_with_clustering_gsm8k_df, reg_with_clustering_cw_df, logit_no_clustering_cw_compliance_df, lpm_with_clustering_cw_compliance_df], ignore_index=True)

# Limit columns to Title, conversation_length_thousands, consolidated_num_steps_ideas, conversation_length_thousands_2, consolidated_num_steps_ideas_2, fres, fres_2
results_df = results_df[['Title', 'conversation_length_thousands', 'conversation_length_thousands_2', 'consolidated_num_steps_ideas', 'consolidated_num_steps_ideas_2', 'fres', 'fres_2']]

# Replace NaN with blanks
results_df = results_df.fillna('')

# Rename Title to Model, rename conversation_length_thousands to Conversation Length, rename consolidated_num_steps_ideas to Number of Steps/Ideas, rename conversation_length_thousands_2 to Conversation Length Squared, rename consolidated_num_steps_ideas_2 to Number of Steps/Ideas Squared, rename fres to FRES, rename fres_2 to FRES Squared
results_df = results_df.rename(columns={'Title': 'Model', 'conversation_length_thousands': 'Conversation Length (Thousands of Tokens)', 'consolidated_num_steps_ideas': 'Number of Steps/Ideas', 'conversation_length_thousands_2': 'Conversation Length (Thousands of Tokens) Squared', 'consolidated_num_steps_ideas_2': 'Number of Steps/Ideas Squared', 'fres': 'Flesch Reading Ease', 'fres_2': 'Flesch Reading Ease Squared'})

# Output to latex. center columns, wrap text, and remove index
latex_string = results_df.to_latex(index=False, 
                      column_format='x{1.5cm}|x{1.5cm}|x{1.5cm}|x{1.5cm}|x{1.5cm}|x{1.5cm}|x{1.5cm}', 
                      #booktabs = True
                      #longtable = True
                      )

# Add lines between rows
lines = latex_string.split('\n')
new_lines = []
for line in lines:
    new_lines.append(line)
    if '\\' in line and '&' in line:  # Identifies a row of the table
        new_lines.append('\\hline')
# Insert \\hline after \toprule
new_lines.insert(2, '\\hline')

# Rejoin the modified lines
modified_latex_table = '\n'.join(new_lines)

print(modified_latex_table)

# Save string to file
with open('../Output for Report/regressions.tex', 'w') as f:
    f.write(modified_latex_table)

# Print results_df
print(results_df)

# Print modifed latex table
print(modified_latex_table)
